IOS-4104 Extracted RemoteStorageSegmentInfo building logic
